[glibc] Don't use rpath-prefix for $ORIGIN
Don't use rpath-prefix for $ORIGIN-prefixed RPATH/RUNPATH. Otherwise,
resolved path will be duplicated, making DSO search to fail. E.g., for
following executable:

  /srv/mer/targets/Sailfish-latest-i486/usr/lib/jvm/java-16-openjdk/bin/java

with RUNPATH set to:

    $ORIGIN:$ORIGIN/../lib

one of the search paths will be the following:

    /srv/mer/targets/Sailfish-latest-i486/srv/mer/targets/Sailfish-latest-i486/usr/lib/jvm/java-16-openjdk/bin/../lib/libjli.so

instead of this (which is valid):

    /srv/mer/targets/Sailfish-latest-i486/usr/lib/jvm/java-16-openjdk/bin/../lib/libjli.so

Excluding rpath-prefix for $ORIGIN-prefixed rpaths eliminates this
problem.